backend/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
distro==1.9.0
Django==5.1.2
django-celery-beat==2.7.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
git-filter-repo==2.45.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.7.0
kombu==5.4.2
openai==1.53.0
packaging==24.1
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
